added drop tables migration.

+++ database/migrations/2025_07_02_202329_drop_reason_column_in_unavailable_dates_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('unavailable_dates', function (Blueprint $table) {
-            $table->dropColumn('reason');
+           if (Schema::hasColumn('unavailable_dates', 'reason')) {
+                $table->dropColumn('reason');
+            }
         });
     }
